+++ modules/robot/src/real-robot/afma4/vpServolens.cpp
@@ -811,7 +811,7 @@ vpServolens::read(char *c, long timeout_s) const
 
   if (select(FD_SETSIZE, &readfds, (fd_set *)NULL,
 	     (fd_set *)NULL, &timeout) > 0) {
-    int n = ::read(this->remfd, c, 1); /* read one character at a time */
+    ssize_t n = ::read(this->remfd, c, 1); /* read one character at a time */
     if (n != 1)
       return false;
     *c &= 0x7f;
